gcp/pam/cloud-functions/pam-request-grant-create/requirements.txt
functions-framework==3.*
google-cloud-privilegedaccessmanager
protobuf
